+++ dashboard/package.json
@@ -22,6 +22,11 @@
     "react": "^15.4.2",
     "react-bootstrap": "^0.30.7",
     "react-dom": "^15.4.2",
+    "react-redux": "^5.0.3",
+    "redux": "^3.6.0",
+    "redux-logger": "^2.8.2",
+    "redux-persist": "^4.4.2",
+    "redux-thunk": "^2.2.0",
     "screenfull": "^3.0.2",
     "uuid": "^3.0.1",
     "vis": "^4.18.1"
