+++ dashboard/src/reducers/index.js
@@ -1,0 +1,10 @@
+import { combineReducers } from "redux";
+import queries from "./queries";
+import response from "./response";
+
+const rootReducer = combineReducers({
+    queries,
+    response,
+});
+
+export default rootReducer;
